Added function documentation
